Extract shared KanjiDMAK + StrokeOrderReplay components
The near-identical dmak animation components in StrokeAnimation and
StrokeOrderPlayer are unified in components/common/KanjiDmak.tsx, using
the KanjiDetails superset (staticMode, gridShow). Both former hosts are
now thin layout wrappers; per-screen styling flows through className
props.
